[#18360] YSQL: fix yb_hash_code compared to an out of range constant
Summary:
The conditions like `yb_hash_code(<key>) <op> <const>`, where <key>
matches hashed part of the primary key or a secondary index are
specially handled. Our optimizer uses the <const> as a scan's
lower or upper hash bound, or both, depending on <op>.

The hash code range is [0..65535], but Postgres allows <const> to be
up to int8, since int8 is directly comparable to int4, the declared
return type of the yb_hash_code function. The PgGate performed
unchecked type cast to hash code's uint16_t data type, causing data
loss if values were out of the range.

To fix this issue we add a range check on the Postgres side. If the
yb_hash_code is compared to an out of bound value, treat the
expression as a constant false, or constant true. Also validate
the values in the PgGate by using narrow_cast, which fails whether
the value is out of the target type's range.
Jira: DB-7345

Test Plan: ybd --java-test org.yb.pgsql.TestPgRegressProc

Reviewers: dmitry, jason, tnayak

Reviewed By: jason

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D36446

+++ src/yb/yql/pggate/ybc_pggate.cc
@@ -200,7 +200,8 @@ inline std::optional<Bound> MakeBound(YBCPgBoundType type, uint64_t value) {
   if (type == YB_YQL_BOUND_INVALID) {
     return std::nullopt;
   }
-  return Bound{.value = value, .is_inclusive = (type == YB_YQL_BOUND_VALID_INCLUSIVE)};
+  return Bound{.value = narrow_cast<uint16_t>(value),
+               .is_inclusive = (type == YB_YQL_BOUND_VALID_INCLUSIVE)};
 }
 
 Status InitPgGateImpl(const YBCPgTypeEntity* data_type_table,
@@ -1331,8 +1332,8 @@ YBCStatus YBCPgDmlBindColumnCondIn(YBCPgStatement handle, YBCPgExpr lhs, int n_a
 
 YBCStatus YBCPgDmlBindHashCodes(
   YBCPgStatement handle,
-  YBCPgBoundType start_type, uint64_t start_value,
-  YBCPgBoundType end_type, uint64_t end_value) {
+  YBCPgBoundType start_type, int start_value,
+  YBCPgBoundType end_type, int end_value) {
   return ToYBCStatus(pgapi->DmlBindHashCode(
       handle, MakeBound(start_type, start_value), MakeBound(end_type, end_value)));
 }
