+++ src/yb/yql/pggate/ybc_pggate.h
@@ -506,8 +506,8 @@ YBCStatus YBCPgDmlBindRow(
 YBCStatus YBCPgDmlGetColumnInfo(YBCPgStatement handle, int attr_num, YBCPgColumnInfo* info);
 
 YBCStatus YBCPgDmlBindHashCodes(YBCPgStatement handle,
-                                YBCPgBoundType start_type, uint64_t start_value,
-                                YBCPgBoundType end_type, uint64_t end_value);
+                                YBCPgBoundType start_type, int start_value,
+                                YBCPgBoundType end_type, int end_value);
 
 // For parallel scan only, limit fetch to specified range of ybctids
 YBCStatus YBCPgDmlBindRange(YBCPgStatement handle,
